Refactor `build` to `dist`, add more test scripts and fix brittle test paths
