+++ .gitignore
@@ -1,6 +1,6 @@
 node_modules
 /npm-*
-/dist
+/build
 .coverage
 .directory
 .DS_STORE*
