+++ jest.config.js
@@ -6,6 +6,7 @@ module.exports = {
   "coverageDirectory": ".coverage",
   "testPathIgnorePatterns": [
     "/node_modules/",
+    "/build/",
     "d.ts"
   ]
 }
