+++ src/__tests__/fixtures/package.json
