+++ src/__tests__/fixtures/tsconfig.json
@@ -6,8 +6,6 @@
     "outDir":"./build",
     "sourceMap": true,
     "declaration": true,
-    "noUnusedLocals": true,
-    "downlevelIteration": true,
     "lib": [
       "es6",
       "es7",
@@ -15,5 +13,7 @@
     ]
   },
   "exclude": [],
-  "include": [ "src/**/*" ]
+  "include": [
+    "src/**/*"
+  ]
 }
