--- test/fixtures/tsconfig.json
@@ -1,26 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "es5",
-    "module": "commonjs",
-    "moduleResolution": "node",
-    "outDir":"./dist",
-    "inlineSourceMap": true,
-    "inlineSources": true,
-    "allowJs": false,
-    "allowSyntheticDefaultImports": true,
-    "declaration": true,
-    "lib": [
-      "es2015",
-      "es2016",
-      "es2017.object"
-    ]
-  },
-  "exclude": [
-    "node_modules",
-    "dist"
-  ],
-  "include": [
-    "src/**/*",
-    "**/*"
-  ]
-}
